backend/requirements.txt
alembic==1.13.2
celery==5.4.0
Flask==3.0.3
Flask_Cors==5.0.0
flask_login==0.6.3
Flask_Migrate==4.0.7
Flask_RESTful==0.3.10
flask_sqlalchemy==3.1.1
openai==1.47.1
python-dotenv==1.0.1
SQLAlchemy==2.0.20
psycopg2-binary
redis